Form 2 and some fixes
I made the model, form and template from the second form, it is connected on the views, so it works on the web page.
The data from the form 1 and form 2 wasn't showing up on the ADMIN page from DJANGO, so i had to edit the admin file from 'protocolos'.
There was a lot of learning in this commit 🥳